+++ src/scribe_data/cli/interactive.py
@@ -16,12 +16,15 @@ from rich.logging import RichHandler
 from rich.table import Table
 from tqdm import tqdm
 
+from scribe_data.cli.convert import convert_wrapper
+
 # from scribe_data.cli.list import list_wrapper
 from scribe_data.cli.get import get_data
 from scribe_data.cli.total import total_wrapper
 from scribe_data.utils import (
     DEFAULT_DUMP_EXPORT_DIR,
     DEFAULT_JSON_EXPORT_DIR,
+    DEFAULT_SQLITE_EXPORT_DIR,
     data_type_metadata,
     language_metadata,
     list_all_languages,
@@ -51,6 +54,9 @@ class ScribeDataConfig:
         self.output_dir: Path = Path(DEFAULT_JSON_EXPORT_DIR)
         self.overwrite: bool = False
         self.configured: bool = False
+        self.identifier_case: str = "camel"
+        self.input_dir: Path = Path(DEFAULT_JSON_EXPORT_DIR)
+        self.output_dir_sqlite: Path = Path(DEFAULT_SQLITE_EXPORT_DIR)
 
 
 config = ScribeDataConfig()
@@ -166,12 +172,16 @@ def configure_settings():
 
     output_type_completer = create_word_completer(["json", "csv", "tsv"])
     config.output_type = prompt(
-        "Select output type (json/csv/tsv): ", completer=output_type_completer
+        "Select output type (json/csv/tsv): ",
+        default="json",
+        completer=output_type_completer,
     )
     while config.output_type not in ["json", "csv", "tsv"]:
         rprint("[yellow]Invalid output type selected. Please try again.[/yellow]")
         config.output_type = prompt(
-            "Select output type (json/csv/tsv): ", completer=output_type_completer
+            "Select output type (json/csv/tsv): ",
+            default="json",
+            completer=output_type_completer,
         )
 
     # MARK: Output Directory
@@ -216,21 +226,22 @@ def run_request():
             for data_type in config.selected_data_types:
                 pbar.set_description(f"Exporting {language} {data_type} data")
 
-                if get_data(
-                    language=language,
-                    data_type=data_type,
-                    output_type=config.output_type,
-                    output_dir=str(config.output_dir),
-                    overwrite=config.overwrite,
-                    interactive=True,
-                ):
-                    logger.info(
-                        f"[green]✔ Exported {language} {data_type} data.[/green]"
+                try:
+                    get_data(
+                        language=language,
+                        data_type=data_type,
+                        output_type=config.output_type,
+                        output_dir=str(config.output_dir),
+                        overwrite=config.overwrite,
+                        interactive=True,
                     )
-
-                else:
+                    # The data was successfully written to file, so we can log success
                     logger.info(
-                        f"[red]✘ Failed to export {language} {data_type} data.[/red]"
+                        f"[green]✔ Successfully exported {language} {data_type} data.[/green]"
+                    )
+                except Exception as e:
+                    logger.error(
+                        f"[red]✖ Failed to export {language} {data_type} data: {str(e)}[/red]"
                     )
 
                 pbar.update(1)
@@ -320,51 +331,123 @@ def start_interactive_mode(operation: str = None):
     """
     while True:
         # Check if both selected_languages and selected_data_types are empty.
-        if not config.selected_languages and not config.selected_data_types:
-            if operation == "get":
-                choices = [
-                    questionary.Choice("Configure get data request", "configure"),
-                    # Choice("See list of languages", "languages"),
-                    questionary.Choice("Exit", "exit"),
-                ]
-
-            elif operation == "total":
-                choices = [
-                    questionary.Choice("Configure total lexemes request", "total"),
-                    # Choice("See list of languages", "languages"),
-                    questionary.Choice("Exit", "exit"),
-                ]
-            elif operation == "translations":
-                choices = [
-                    questionary.Choice(
-                        "Configure translations request", "translations"
-                    ),
-                    # Choice("See list of languages", "languages"),
-                    questionary.Choice("Exit", "exit"),
-                ]
-
-        else:
+        if config.selected_languages or config.selected_data_types:
             choices = [
                 questionary.Choice("Configure get data request", "configure"),
                 questionary.Choice("Exit", "exit"),
             ]
             if config.configured:
-                choices.insert(1, questionary.Choice("Request for get data", "run"))
+                choices.insert(
+                    1, questionary.Choice("Run get data request with WDQS", "run")
+                )
+                choices.insert(
+                    2,
+                    questionary.Choice(
+                        "Run get lexemes request with lexeme dumps", "run_all"
+                    ),
+                )
 
+            elif config.selected_languages and config.selected_data_types:
+                choices.insert(
+                    1, questionary.Choice("Request for convert JSON", "convert_json")
+                )
             else:
                 choices.insert(
                     1, questionary.Choice("Request for total lexeme", "total")
                 )
 
+        elif operation == "get":
+            choices = [
+                questionary.Choice("Configure get data request", "configure"),
+                # Choice("See list of languages", "languages"),
+                questionary.Choice("Exit", "exit"),
+            ]
+
+        elif operation == "total":
+            choices = [
+                questionary.Choice("Configure total lexemes request", "total"),
+                # Choice("See list of languages", "languages"),
+                questionary.Choice("Exit", "exit"),
+            ]
+        elif operation == "convert":
+            choices = [
+                questionary.Choice("Configure convert request", "convert"),
+                questionary.Choice("Exit", "exit"),
+            ]
+        elif operation == "translations":
+            choices = [
+                questionary.Choice("Configure translations request", "translations"),
+                # Choice("See list of languages", "languages"),
+                questionary.Choice("Exit", "exit"),
+            ]
+
         choice = questionary.select("What would you like to do?", choices=choices).ask()
 
         if choice == "configure":
             configure_settings()
+        elif choice == "run_all":
+            if wikidata_dump_path := prompt(
+                f"Enter Wikidata lexeme dump path (default: {DEFAULT_DUMP_EXPORT_DIR}): "
+            ):
+                wikidata_dump_path = Path(wikidata_dump_path)
+            else:
+                wikidata_dump_path = Path(DEFAULT_DUMP_EXPORT_DIR)
 
+            parse_wd_lexeme_dump(
+                language=config.selected_languages,
+                wikidata_dump_type=["get"],
+                data_types=config.selected_data_types,
+                type_output_dir=config.output_dir,
+                wikidata_dump_path=wikidata_dump_path,
+            )
+            rprint(THANK_YOU_MESSAGE)
+            break
         elif choice == "total":
             prompt_for_languages()
             prompt_for_data_types()
             request_total_lexeme_loop()
+            break
+
+        elif choice == "convert":
+            prompt_for_languages()
+            prompt_for_data_types()
+
+            # Use the default explicitly so that if the user enters nothing, the default value is retained.
+            user_input_dir = prompt(
+                f"Enter input directory (default: {config.input_dir}): ",
+                default=str(config.input_dir),
+            )
+            config.input_dir = Path(
+                user_input_dir
+            )  # Update the configuration with the user's input.
+            user_output_dir = prompt(
+                f"Enter output directory (default: {config.output_dir_sqlite}): ",
+                default=str(config.output_dir_sqlite),
+            )
+            config.output_dir_sqlite = Path(user_output_dir)
+            identifier_case = prompt(
+                "Enter identifier case (default: camel): ",
+                default="camel",
+            )
+            output_type = prompt(
+                "Enter output type (default: sqlite): ",
+                default="sqlite",
+            )
+            overwrite_str = prompt(
+                "Overwrite existing files? (default: False): ",
+                default="False",
+            )
+            overwrite_bool = overwrite_str.strip().lower() in ("true", "y", "yes")
+
+            convert_wrapper(
+                languages=config.selected_languages,
+                data_types=config.selected_data_types,
+                output_type=output_type,
+                input_files=config.input_dir,  # Use the updated configuration value.
+                output_dir=config.output_dir_sqlite,
+                identifier_case=identifier_case,
+                overwrite=overwrite_bool,
+            )
             break
 
         elif choice == "translations":
@@ -389,10 +472,6 @@ def start_interactive_mode(operation: str = None):
             )
 
             break
-
-        # elif choice == "languages":
-        #     see_list_languages()
-        #     break
 
         elif choice == "run":
             run_request()